backend/requirements.txt
fastapi
uvicorn
pydantic
requests
reportlab>=4.0.0
python-dotenv
